requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiomysql>=0.2.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
itsdangerous>=2.1.2
httpx[http2]>=0.27.0
sse-starlette>=2.1.0
ldap3>=2.9.1
pypdf>=4.2.0
python-docx>=1.1.0
python-pptx>=0.6.23
apscheduler>=3.10.4
pydantic-settings>=2.3.0
fpdf2>=2.7.9
redis>=5.0.0
